Fix CRC32 hashing for x86_64

+++ neo/idlib/Dict.cpp
@@ -235,7 +235,7 @@ idDict::Checksum
 ================
 */
 int	idDict::Checksum( void ) const {
-	unsigned long ret;
+	unsigned int ret;
 	int i, n;
 
 	idList<idKeyValue> sorted = args;
